requirements.txt
astroid==2.4.1
cachelib==0.1.1
certifi==2020.6.20
cssselect==1.1.0
EasyProcess==0.3
entrypoint2==0.2.1
howdoi==1.2.1
isort==4.3.21
jeepney==0.4.3
lazy-object-proxy==1.4.3
lxml==4.5.1
mccabe==0.6.1
MouseInfo==0.1.3
mss==6.0.0
numpy==1.19.2
opencv-python==4.4.0.42
Pillow==7.2.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
Pygments==2.6.1
pylint==2.5.2
PyMsgBox==1.0.8
pyperclip==1.8.0
pyquery==1.4.1
PyRect==0.1.4
pyscreenshot==2.2
PyScreeze==0.1.26
python3-xlib==0.15
PyTweening==1.0.3
pytz==2020.1
tesserocr==2.5.1
wpm==1.51.5
wrapt==1.12.1
